# Conflicts:
#	app/Http/Controllers/FileController.php
#	public/css/login.css

+++ app/Http/Controllers/FileController.php
@@ -12,12 +12,37 @@ use Illuminate\Support\Facades\Validator;
 
 class FileController extends Controller
 {
+
+    public function logoutTime(){
+
+        $curr = time();
+        $last = Session::get('time');
+        if($curr - $last > 10){
+            return redirect('/logout');
+        }
+        Session::put('time', time());
+
+    }
+
     public function index(){
+        $curr = time();
+        $last = Session::get('time');
+        if($curr - $last > 10){
+            return redirect('/logout');
+        }
+        Session::put('time', time());
         $files = DB::table('files')->where('TraineeCode', 'like', Auth::user()->TraineeCode)->simplePaginate(4);
         return view('homepage', compact('files'));
     }
 
     public function insert(Request $request){
+        $curr = time();
+        $last = Session::get('time');
+        if($curr - $last > 10){
+            return redirect('/logout');
+        }
+        Session::put('time', time());
+
         $rules = [
             'file' => 'required'
         ];
@@ -43,6 +68,13 @@ class FileController extends Controller
     }
 
     public function delete($id){
+        $curr = time();
+        $last = Session::get('time');
+        if($curr - $last > 10){
+            return redirect('/logout');
+        }
+        Session::put('time', time());
+
         $file = File::find($id);
         if(isset($file)){
             Storage::delete('public/files/'.$file->discriminator);
@@ -52,6 +84,13 @@ class FileController extends Controller
     }
 
     public function download($id){
+        $curr = time();
+        $last = Session::get('time');
+        if($curr - $last > 10){
+            return redirect('/logout');
+        }
+        Session::put('time', time());
+
         $file = File::find($id);
         return Storage::download('public/files/'.$file->name);
     }
